Organize files and setup boilerplate components.
