+++ src/card/Card.tsx
@@ -1,0 +1,12 @@
+import './Card.css';
+
+function Card() {
+
+    return (
+        <>
+        
+        </>
+    )
+}
+
+export default Card;
