+++ src/home/Home.tsx
@@ -1,0 +1,13 @@
+import './Home.css';
+import Main from '../main/Main';
+
+function Home() {
+
+    return (
+        <>
+        
+        </>
+    )
+}
+
+export default Home;
