+++ src/main/Main.tsx
@@ -1,0 +1,12 @@
+import './Main.css';
+
+function Main() {
+
+    return (
+        <>
+        
+        </>
+    )
+}
+
+export default Main;
