+++ src/nav/Nav.tsx
@@ -1,0 +1,12 @@
+import './Nav.css';
+
+function Nav() {
+
+    return (
+        <>
+        
+        </>
+    )
+}
+
+export default Nav;
