+++ src/teas/Teas.tsx
@@ -1,0 +1,13 @@
+import './Teas.css';
+import Card from '../card/Card';
+
+function Teas() {
+
+    return (
+        <>
+        
+        </>
+    )
+}
+
+export default Teas;
